Properly escape all HTML output
We should not assume that some values passed to templates don't need to
be escaped; it's just to error prone.  Instead we recursively escape
all values in the passed `$bag`.  This also prevents objects to be
passed to the templates, so we use "DTOs" (well, actually just arrays)
instead, what might be easier to grasp for designers and webmasters.

This is quite different from the approach chosen for (early versions
of) Pfw_XH – both have their pros and cons.