Close #1: Load images on demand
This is the most simplistic implementation we can think of, but it also
might the most effective: we use the `loading` attribute of the `img`
element to load the first slide show image `eager`, but all others
`lazy`.  This should be sufficient to cover all of the alternative
`picture` `source`s.  Of course, the effect heavily depends on browser
support.